+++ Code/Main Project/CycleBell/ViewModels/TimePointViewModels/BeginTimePointViewModel.cs
@@ -23,7 +23,7 @@ namespace CycleBell.ViewModels.TimePointViewModels
 {
     public class BeginTimePointViewModel : TimePointViewModelBase
     {
-        public BeginTimePointViewModel(byte loopNumber, IPresetViewModel presetViewModel) : base((int)TimePoint.MinId, loopNumber, presetViewModel)
+        public BeginTimePointViewModel(int loopNumber, IPresetViewModel presetViewModel) : base(TimePoint.MinId, loopNumber, presetViewModel)
         { }
 
         public override TimePoint TimePoint => null;
